Revert "use more defaults and simplify our options"
This reverts commit 051f7e4dc249f87fc1eea92dbba3a20f0043ed34, 

Reason for revert: OVB jobs in CI is failing during 
"Manage firewall rules" task because the "999 drop all"
rule gets applied first. Details in [1]


We suspect the above commit caused performance issues and
on slow connections where ansible may need to reconnect
over SSH while the firewall rules are being applied,
the connection could be denied if the allow rule for
SSH is not yet added. The deployment will then hang
and fail.


[1] https://bugs.launchpad.net/tripleo/+bug/1969158

Change-Id: Ia8e508254b578b35d3e3aee773e183025b1a9cef