Fix unused DEFAULT_CONTAINER_TAG
The DEFAULT_CONTAINER_TAG constant is unused but the hard-coded value
is used instead of the constants. This change ensures the constant is
used to gather all default values in a single place.

Also, this replaces the hard-coded name prefix by the constant so that
we all define default values for container image detection as
constants.

Change-Id: I4a3b2c2943680afde67f13daefd7aaf8258132cc