Remove duplicate check from tripleo-deploy
The much debated and controversial duplicate parameter check
is a little too noisy for the undercloud install. We include
a lot of environment files when we run the undercloud install /
upgrade. This creates a lot of unnecessary noise that will
probably overwhelm users.

Let's just restrict it to overcloud operations which is the
primary place that I believe people wanted to see this feature.

Change-Id: Ie00e4517817bdb53573193e07b3423df89f716d1