Fall back to alt path for ceph client data during export
The patch to fix the related bug (not this patch) fixes an
inconsistency where the ceph client data is somtimes exported
to the wrong path. If anyone deployed before they got the fix
to the related bug, then this patch will workaround that
problem if they export later.

When 'openstack overcloud export ceph' is run look in the
correct default location for the ceph client data and if it
is not found, then fall back to incorrect location.

Change-Id: I0b8c747481e561645eefea50cb77ca806a7f0de2
Related-Bug: #1978846