Refactor deploy validations for ephemeral heat
All these validations are not executed with ephemeral
heat. With this we'll save stack environment to disk
and validate using it in the next deployment.

Note: Some of these validations have been changed to
not use saved stack environment as we don't update
the stack anymore.

Also cleans up some usage of existing stack in code.

Change-Id: I791407490dcdddb927d488aa73856cd3ef7191bb