Stop using heat-admin as ansible_ssh_user
We use tripleo-admin when running config-downloaded
playbooks with overcloud deploy, update and upgrade.
We used to generate the inventory with tripleo-admin
as the default ansible_ssh_user. After we moved to
generate inventory with node provision the default
ansible_ssh_user has changed to heat-admin. This
patch drops heat-admin user completely as we don't
need it anymore.

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart/+/849807
Change-Id: I51c3d6af40b6b268bb8e3fb26fe4c2833254549f