Support remove and revert actions
The CLI should not only support creating snapshots, but it also
must support two other actions: remove and revert. Remove deletes
all the snapshots that were created in a previous execution,
while revert loads the information stored in the snapshots to the
main disks, leaving them in the same state they were when the
snapshots were created.

Change-Id: I75925c749acd26cbdf62991e7e0ec87d906d5def