Use realpath for checking directories
This patch changes the method we use to check
for existing paths to use the canonical path
rather that the literal value. This is really
more of a quality of life change that makes
unit tests work on multiple OS versions.

Namely, it makes the unittests work on MacOS.
But should also be applicable to Windows should
someone want to contribute from a Windows or
MacOS based platform.

Change-Id: Ic02839f14b7ff60d02aeb8486bbf89f78c0d1dd5