Support role specific ContainerImagePrepare
This is the last change of the series of changes to implements partial
support for overriding the ContainerImaePrepare parameter for
individual roles, so that users can use different set of images for
each role.

example:
parameter_defaults:
  ContainerImagePrepare:
    - push_destination: true
      set:
        name_prefix: openstack-
        name_suffix: ''
        namespace: quay.io/repository/tripleowallabycentos9
        ...
  ComputeEl8ContainerImagePrepare:
    - push_destination: true
      set:
        name_prefix: openstack-
        name_suffix: ''
        namespace: quay.io/repository/tripleowallabycentos8
        ...

This change makes the deployment command to generate the image
parameters in tht according to these layered definitions.

Depends-on: https://review.opendev.org/851690
Change-Id: I76f78a12f6b4b0763a47e85f38f743da6269a47e