Add cell export to overcloud deploy
Run the equivalent of cell export after the stack is deployed.
The difference between overcloud export and cell export is that
should_filter=True when running cell export. As either of these exports
might be needed later to deploy other stacks, go ahead and make sure
both are always run.

Change-Id: Idceeb31609533b78034f0b25c4b0fbc6d47c19fb
Signed-off-by: James Slagle <jslagle@redhat.com>