Remove migration from undercloud-passwords.conf
This removes the logic to look up the existing passwords from
the undercloud-passwords.conf file during upgrade.

The file was generated by instackenv which is no longer used since
Rocky, thus the migration should have been done when a deployment is
upgraded from Queens to Rocky(Upgrade) or Train(FFU), and the logic
is no longer required in recent releases.

Change-Id: Id0aa63e4909aac5309c3828d0334cbeca08f5b30