Do not allow uc upgrade when network plugin changes
If the undercloud is upgraded with a change in network plugin
i.e ovs to ovn or ovn to ovs it will break the undercloud as
just switching is not enough it needs network resources to be
migrated, so we detect if there is change in network and block
the upgrade.

[1] is switching default network plugin on Undercloud
from ovs to ovn. But as direct upgrade is not possible
from ovs to ovn adding a check for the same and block
the undercloud upgrade.

[1] https://review.opendev.org/850158

Change-Id: I4a80380ab28edf318019cb0bf05765d329ed0405