Make ~/.tripleo/history owned by the correct user
When a tripleo subcommand command is executed with sudo, we identify
the original user and create the history file in that user's home
directory.

When creating the file, we should make that user owns it, otherwise
the user no longer able to run tripleo subcommends without sudo once
the history file is created and owned by root.

Change-Id: Ie6195b629fa65bcf9e5350f73d332e5034e8039b