parse_ansible_inventory: use ansible_runner instead of importing Ansible
All our interactions with Ansible are brokered by the ansible_runner
module, except for this one function which imports the Ansible module
proper. This leads for breakage when using the latest Ansible builds
from EL9, which are on python3.11 instead of 3.9.

By refactoring parse_ansible_inventory to use ansible_runner's
get_inventory we avoid that breakage, but we also lose out on the
pattern matching as that isn't supported. This function is called in two
places:

ceph_hosts_in_inventory: it didn't do any pattern matching at all, so it
isn't impacted by the change.

overcloud admin authorize: we can now offload the pattern matching to
Ansible itself via the limit_hosts parameter we pass to the playbook run
command, preserving the original functionality.

Closes-Bug: #2007659
Change-Id: I08330ced3b31d0ff48a881501cae8bd726e36432