Fix Ceph Stray host(s)/daemon(s) issue
During deployed ceph, spec is created with short hostnames and it is
applied to create all the ceph deamons on respective hosts.

In this stage if TLS is configured, hostnames are modified to
canonical names (fqdn), which results in stray ceph host(s)/daemon(s).

This patch adds the new option '--tld' to 'openstack overcloud ceph spec'
and 'openstack overcloud ceph deploy' commands.

When tld is passed to 'ceph_spec_bootstrap' module, ceph spec generation
is done with updated hostnames.
Also the hostnames are updated based on the tld passed.

Resolves: rhbz#2173101
Change-Id: Ib67785adcdd12892ff9038c21a3582746930b400