Fix node extract when no network-isolation
When network isolation is not enabled by
environments/network-isolation.yaml the networks in the default
networks file is not created. But the RoleNetIpMap still have
entries for each network all with the ctlplane IP address.

This change re-factors the code to ignore RoleNetIpMap entries
when the network is not found/does not exist.

Change-Id: If35ff8a1b3d9b4e93e6151c9c28d8e9c707bc64a
Closes-Bug: #2027580